Refactor package of AsymmetricTextEmbeddingParameters
The MLCommonsClassLoader expects all MLAlgoParameters to be in the
"org.opensearch.ml.common.input.parameter" package.
