+++ common/src/test/java/org/opensearch/ml/common/dataset/AsymmetricTextEmbeddingParametersTest.java
@@ -11,7 +11,8 @@ import org.opensearch.ml.common.TestHelper;
 
 import java.io.IOException;
 import java.util.function.Function;
-import org.opensearch.ml.common.dataset.AsymmetricTextEmbeddingParameters.EmbeddingContentType;
+import org.opensearch.ml.common.input.parameter.textembedding.AsymmetricTextEmbeddingParameters;
+import org.opensearch.ml.common.input.parameter.textembedding.AsymmetricTextEmbeddingParameters.EmbeddingContentType;
 
 import static org.junit.Assert.assertEquals;
 import static org.opensearch.ml.common.TestHelper.contentObjectToString;
